fix(vault): move "Move to section…" into the file overflow menu
Per review: the move action belongs in the image's "More options" overflow menu
(next to Share / Delete), not as a row in the bottom detail sheet. Add a
"Move to section…" item to VaultFileDropdownMenu that opens a section sub-menu
(other sections only); remove MoveToSectionRow and its plumbing from
VaultGalleryDetailSheet.
